Fix symbol table bug on 2.13 caused by `scala.Seq`.
Previously, we matched against `case scala.Seq` in metacp when loading
symbols from a classpath. This caused a match error in 2.13 at runtime
since we deconstructed a `mutable.Buffer`. This bug was not caught in
tests because it turns out we only ran symbol table tests in 2.12.

This commit both fixes the bug and enables the symbol table tests
for all Scala versions (2.11, 2.12, 2.13).
