--- tests/jvm/src/main/scala-2.12/scala/meta/tests/metacp/Jars.scala
@@ -1,94 +1,0 @@
-package scala.meta.tests.metacp
-
-import coursier._
-import coursier.util.Gather
-import coursier.util.Task
-import java.io.OutputStreamWriter
-import java.io.PrintStream
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.meta.AbsolutePath
-
-case class ModuleID(organization: String, name: String, version: String) {
-  def toCoursier: Dependency = Dependency(Module(organization, name), version)
-  override def toString: String = s"$organization:$name:$version"
-}
-object ModuleID {
-  def scalaReflect(scalaVersion: String): ModuleID =
-    ModuleID("org.scala-lang", "scala-reflect", scalaVersion)
-  def fromString(string: String): List[ModuleID] = {
-    string
-      .split(";")
-      .iterator
-      .flatMap { moduleId =>
-        moduleId.split(":") match {
-          case Array(org, name, rev) =>
-            ModuleID(org, name, rev) :: Nil
-          case _ => Nil
-        }
-      }
-      .toList
-  }
-}
-object Jars {
-  def fetch(
-      org: String,
-      artifact: String,
-      version: String,
-      out: PrintStream = System.out,
-      // If true, fetches the -sources.jar files instead of regular jar with classfiles.
-      fetchSourceJars: Boolean = false,
-      provided: List[ModuleID] = Nil
-  ): List[AbsolutePath] = {
-    fetch(ModuleID(org, artifact, version) :: provided, out, fetchSourceJars)
-  }
-
-  def fetch(
-      modules: Iterable[ModuleID],
-      out: PrintStream,
-      fetchSourceJars: Boolean
-  ): List[AbsolutePath] = {
-    val classifier = if (fetchSourceJars) "sources" :: Nil else Nil
-    val res = Resolution(modules.toIterator.map(_.toCoursier).toSet)
-    val repositories = Seq(
-      Cache.ivy2Local,
-      MavenRepository("https://repo1.maven.org/maven2"),
-      // For scala-compiler nightly builds
-      MavenRepository("https://scala-ci.typesafe.com/artifactory/scala-integration/")
-    )
-    val term =
-      new TermDisplay(new OutputStreamWriter(out), fallbackMode = true)
-    term.init()
-    val fetch = Fetch.from(repositories, Cache.fetch[Task](logger = Some(term)))
-    val resolution = res.process.run(fetch).unsafeRun()
-    val errors = resolution.errors
-    if (errors.nonEmpty) {
-      sys.error(errors.mkString("\n"))
-    }
-    val artifacts: Seq[Artifact] =
-      if (fetchSourceJars) {
-        resolution
-          .dependencyClassifiersArtifacts(classifier)
-          .map(_._2)
-      } else resolution.artifacts
-    val localArtifacts = Gather[Task]
-      .gather(
-        artifacts.map(artifact => Cache.file[Task](artifact).run)
-      )
-      .unsafeRun()
-    val jars = localArtifacts.flatMap {
-      case Left(e) =>
-        if (fetchSourceJars) {
-          // There is no need to fail fast here if we are fetching source jars.
-          println(e.describe)
-          Nil
-        } else {
-          throw new IllegalArgumentException(e.describe)
-        }
-      case Right(jar) if jar.getName.endsWith(".jar") =>
-        AbsolutePath(jar) :: Nil
-      case _ => Nil
-    }
-    term.stop()
-    jars.toList
-  }
-}
