+++ tests/jvm/src/main/scala/scala/meta/tests/metacp/Libraries.scala
