+++ tests/jvm/src/test/scala/scala/meta/tests/symtab/SymbolTableSuite.scala
